[Build] Use Borland's real-mode MAKER.EXE
DOSBox-X seems to have some problems with it, or maybe that's just me
having configured something wrong. In any case, it's not like we need
that extra memory, so removing the need for the DPMI driver removes
one more potential source of problems.

Oh, and it has always been part of the devkit. :P