[Decompilation] Define generic function pointer types
Going with the classic pointer-in-typedef approach here, because the
syntax you'd otherwise have to use is terribly inconsistent. It'd be

	 farfunc_t *near near_ptr_to_far_func;

but

	nearfunc_t near *near_ptr_to_near_func;

And that'd hopefully be the last change to ReC98.h for a long time!
Those glacial compile times if every .c file is affected… Really
stands out if your build system is otherwise perfect.

Part of P0030, funded by zorg.