[Build] 32-bit: Actually create `bin/Pipeline` during the 32-bit part -.-
Lol, this led the bmp2arr compilation to fail on a clean checkout.
Lesson learned: Always test on a clean clone, not in your separate test
repo that you just `git pull` up to the final commit in a push, but
with all the build artifacts still in there…

Also, yes, `mkdir` only supports backslashes.

Part of P0001, funded by GhostPhanom.